diagram/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aio-pika=6.8.2=pypi_0
aiormq=3.3.1=pypi_0
atk-1.0=2.36.0=ha1a6a79_0
attrs=21.4.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
ca-certificates=2022.07.19=h06a4308_0
cachetools=5.2.0=pypi_0
cairo=1.16.0=h19f5f5c_2
certifi=2022.6.15=py39h06a4308_0
charset-normalizer=2.0.12=pypi_0
diagrams=0.21.1=pypi_0
et-xmlfile=1.1.0=pypi_0
ex-template-model-gx-adapter=0.0.145=pypi_0
expat=2.4.4=h295c915_0
font-ttf-dejavu-sans-mono=2.37=hd3eb1b0_0
font-ttf-inconsolata=2.001=hcb22688_0
font-ttf-source-code-pro=2.030=hd3eb1b0_0
font-ttf-ubuntu=0.83=h8b1ccd4_0
fontconfig=2.13.1=h6c09931_0
fonts-anaconda=1=h8fa9717_0
fonts-conda-ecosystem=1=hd3eb1b0_0
freetype=2.11.0=h70c0345_0
fribidi=1.0.10=h7b6447c_0
gdk-pixbuf=2.42.8=h433bba3_0
glib=2.69.1=h4ff587b_1
gobject-introspection=1.72.0=py39hbb6d50b_0
google=3.0.0=pypi_0
google-api-core=2.8.2=pypi_0
google-auth=2.9.1=pypi_0
google-cloud=0.34.0=pypi_0
google-cloud-vision=3.0.0=pypi_0
googleapis-common-protos=1.56.4=pypi_0
graphite2=1.3.14=h295c915_1
graphviz=2.50.0=h3cd0ef9_0
grpcio=1.47.0=pypi_0
grpcio-status=1.47.0=pypi_0
grpcio-tools=1.38.1=pypi_0
gtk2=2.24.33=h73c1081_2
gts=0.7.6=hb67d8dd_3
harfbuzz=4.3.0=hd55b92a_0
icu=58.2=he6710b0_3
idna=3.3=pypi_0
iniconfig=1.1.1=pypi_0
jinja2=3.1.2=pypi_0
jpeg=9e=h7f8727e_0
kubernetes=18.20.0=pypi_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.8=h7f8727e_5
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgd=2.3.3=h695aa2c_1
libgomp=11.2.0=h1234567_1
libpng=1.6.37=hbc83047_0
librsvg=2.54.4=h19fe530_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.4.0=hecacb30_0
libtool=2.4.6=h295c915_1008
libuuid=1.0.3=h7f8727e_2
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.15=h7f8727e_0
libxml2=2.9.14=h74e7548_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.1.1=pypi_0
multidict=6.0.2=pypi_0
mypy-protobuf=2.5=pypi_0
ncurses=6.3=h5eee18b_3
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numpy=1.20.3=pypi_0
oauthlib=3.2.0=pypi_0
openpyxl=3.0.10=pypi_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pypi_0
pamqp=2.3.0=pypi_0
pandas=1.2.5=pypi_0
pango=1.50.7=h05da053_0
pcre=8.45=h295c915_0
pip=22.1.2=py39h06a4308_0
pixman=0.40.0=h7f8727e_1
pluggy=1.0.0=pypi_0
prometheus-client=0.11.0=pypi_0
proto-plus=1.20.6=pypi_0
protobuf=3.20.1=pypi_0
psutil=5.9.1=pypi_0
py=1.11.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=3.0.9=pypi_0
pytest=7.1.2=pypi_0
python=3.9.12=h12debd9_1
python-dateutil=2.8.2=pypi_0
python-dotenv=0.20.0=pypi_0
python-graphviz=0.19.2=pypi_0
pytz=2021.3=pypi_0
pyyaml=6.0=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.26.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
setuptools=61.2.0=py39h06a4308_0
six=1.16.0=pypi_0
sortedcollections=2.1.0=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlite=3.38.5=hc218d9a_0
sseclient=0.0.27=pypi_0
th2-check2-recon=3.2.0=pypi_0
th2-common=3.8.1=pypi_0
th2-common-utils=1.1.0=pypi_0
th2-grpc-act-default-le=3.2.28=pypi_0
th2-grpc-act-p8-le=0.2.19=pypi_0
th2-grpc-act-template=3.7.0=pypi_0
th2-grpc-check1=3.5.0=pypi_0
th2-grpc-common=3.8.0=pypi_0
th2-grpc-crawler-data-processor=0.3.1=pypi_0
th2-grpc-data-provider=0.1.6=pypi_0
th2-grpc-util=3.0.4=pypi_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=pypi_0
tqdm=4.64.0=pypi_0
types-protobuf=3.19.22=pypi_0
tzdata=2022a=hda174b7_0
urllib3=1.26.10=pypi_0
uvloop=0.16.0=pypi_0
websocket-client=1.3.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xlrd=1.2.0=pypi_0
xz=5.2.5=h7f8727e_1
yarl=1.7.2=pypi_0
zlib=1.2.12=h7f8727e_2
zstd=1.5.2=ha4553b6_0
